third_party/WebKit/LayoutTests/fast/events/touch/touch-event-source-device-event-sender-expected.txt

Test that touch events should have a sourceDevice and its firesTouchEvents should be true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Sending touchstart
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
Sending touchmove
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
Sending touchend
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
Sending touchstart
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
Sending touchcancel
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
PASS successfullyParsed is true

TEST COMPLETE
